backend/executor-service/requirements.txt
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
idna==3.10
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
minio==7.2.12
msgpack==1.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
protobuf==5.29.1
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
ray==2.40.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xgboost==2.1.0
